Added comment + minor change

+++ lib/main.dart
@@ -7,11 +7,21 @@ void main() async {
   runApp(MyApp(highscore: highscore));
 }
 
+// Cette fonction va aller chercher la valeur highscore
+// qu'on a stockée et va nous la retourné
 Future<int> getHighscore() async {
+  // Ceci est nécessaire si on veut executer du code avant d'utiliser runApp()
   WidgetsFlutterBinding.ensureInitialized();
+
+  // Va créer une instance de SharedPreferences pour que l'on puisse
+  // prendre nos valeurs stockées
   final prefs = await SharedPreferences.getInstance();
+
+  // Va chercher la valeur que l'on souhaite récupérer.
+  // Elle nous retournera 0 si elle ne trouve rien.
   final savedHighscore = prefs.getInt('highscore') ?? 0;
-  return savedHighscore as int;
+
+  return savedHighscore;
 }
 
 class MyApp extends StatelessWidget {
